walk_dict_to_hdf5: Explicitly use np.int64
`dtype=float` is equivalent to `dtype=np.float64`, but `dtype=int` is
equivalent to `dtype=np.int_`, which is 32 bits wide on 32-bit systems
and not necessarily wide enough for the data we're trying to convert.
As shown (sort of) in https://bugs.debian.org/1103147, this caused test
failures on 32-bit architectures.